Refactor precomputation logic
